requirements.txt
dialogflow==0.8.0
python-dotenv==0.12.0
python-telegram-bot==11.1.0
requests==2.22.0
google-cloud-core==1.3.0
vk_api==11.8.0